Editor console warnings: zero the compile-warning set
- DetHash goes internal: the template benchmark assembly references the
  generated Benchmarks assembly, so the copy-through shared class was
  imported on top of its own source (201x CS0436).
- DetMath template (and its tests/benchmark) get pragmas: UseNative is a
  const route so one arm is always dead (CS0162), and x != x is the NaN
  test (CS1718).
- WideOP template: pragma for the template-only dead second dtype arm.
- QR/QRCP: drop three unused QR_BLOCK/QRCP_BLOCK method-local consts.
- AMG Rows becomes a readonly member (CS8656 defensive-copy warning).
- RandomOP int-range guard compares through long, so narrow dtypes stop
  warning CS0652 while the check stays live for long.
- proxy structs get their Equals overrides back (CS0660).
- BlockGCRODRTests: the choose[true|false] advantage assert routes
  through a local so the double variant stops warning CS0162.
- Delete the empty BurstLinearAlgebra.Editor asmdef (per-compile editor
  warning; nothing referenced it).

Suite 7412/7412; fresh headless compile log has zero CS warnings.

+++ Assets/LinearAlgebra/CodeGen/TemplateSource/OP/RandomOP.iProxy.cs
@@ -33,7 +33,7 @@ namespace BULA
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static void nextUniformInPlace(ref Random rng, ref iProxyN dest, iProxy min, iProxy max)
         {
-            if (min < int.MinValue || max > int.MaxValue)
+            if ((long)min < int.MinValue || (long)max > int.MaxValue)
                 throw new ArgumentException("Rand: min/max must be within int range for random generation");
             if (!(min <= max))
                 throw new ArgumentException("Rand: min must be <= max");
@@ -59,7 +59,7 @@ namespace BULA
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static void nextUniformInPlace(ref Random rng, ref iProxyMxN dest, iProxy min, iProxy max)
         {
-            if (min < int.MinValue || max > int.MaxValue)
+            if ((long)min < int.MinValue || (long)max > int.MaxValue)
                 throw new ArgumentException("Rand: min/max must be within int range for random generation");
             if (!(min <= max))
                 throw new ArgumentException("Rand: min must be <= max");
